Avoid compilation failures due to imprecision in memory availability readout
In order to avoid out-of-memory scenarios, the JIT reads the amount of
available physical memory before each compilation. To control the overhead
of the memory availability call, there is a caching mechanism that makes sure
that two such calls cannot be closer than 500ms.
When a compilation consumes most of the available memory for the JVM, it may fail,
and this is completely normal. However, many of the subsequent compilations
can also fail, thinking that there is very little memory available. In reality,
the memory expensive compilation that failed already released all its memory to the
OS, and the JIT is getting stale memory availability information from the caching
mechanism.
The solution is to bypass the caching mechanism and do a proper memory readout,
if the JIT is about to fail a compilation due to insufficient physical memory.

Signed-off-by: Marius Pirvu <mpirvu@ca.ibm.com>