Add extension classpath handling for FIPS 140-3
Adds a parameter to addExtDir() to select which FIPS implementation
should be placed on the extension class path dirs list. This
parameter is used to place either "fips140-2/lib/ext" or
"fips140-3/lib/ext" as the first directory to add to the
extension classpath dirs list. Regardless, the usual default
extension directory ("jre/lib/ext") is also added to the extension
classpath dirs.

A new flag in the J9SpecialArguments structure is used to record
whether the -Xfips140-3 command-line option is active. When
addExtDir() is called, the additional parameter is determined
by the setting oof this new flag.

Following this commit, the java.ext.dirs system property is always
expected to have two directories in it: first a FIPS specific
directory and then the usual default jre/lib/ext directory.

Signed-off-by: Mark Stoodley <mstoodle@ca.ibm.com>