Load JITServer AOT cache file in the background
JITServer offers support for loading an AOT cache from a file.
This is done by a compilation thread that works concurrently with other
compilation threads that compile methods.
The sequence of actions is the following:
1. Compilation thread at the server reads compilation request from client
and determines whether the client wants to use the AOTCache
2. Server checks its hashmap to see if the named cache already exists
If it exists, it will be used.
3. If desired cache does not exist, server checks if the persistence feature
is enabled and whether a file for the desired cache exists.
4. If the AOT cache file exists, the server will memorize the name of the
cache in a set, queue a special compilation request and
wake a compilation thread to process it. It will then proceed to compile
the requestedn method, without waiting for the AOT cache to be loaded.
5. If the AOT cache file does not exist (or persistence is disabled),
the server will create an empty AOT cache which is going to be populated
as methods get compiled.
6. When a compilation thread extracts the special request from the queue,
it will proceed to load the AOT cache from disk. The name of the AOT cache
is retrieved from the set where it was memorized earlier. If there is a
problem during the loading, the name of the cache will be included into a
set of cached that are excluded from being loaded in the future.

Signed-off-by: Marius Pirvu <mpirvu@ca.ibm.com>