Use profilingOpts strategy under JitProfiling mode only
The profilingOpts optimization strategy includes profileGenerator and
deadTreesElimination.  The profileGenerator simply returns if the
JProfiling mode is enabled, while the deadTreesElimination is intended to
clean up trees that might have been introduced by profileGenerator, so it
too is really only needed if JitProfilng mode is enabled.  However, the
optimization strategy as a whole is being run if either JProfiling or
JitProfiling is enabled.

Some optimization strategies run compactLocals very late, just before
profileOpts is run.  Running deadTreesElimination after compactLocals
might invalidate the analysis compactLocals has performed.  The
compactLocals analysis is only run if JitProfiling mode is not
enabled.

This changes uses of profilingGroup to only take effect if
JitProfilng mode is enabled, preventing deadTreesElimination from running
after compactLocals.  It also renames profilingGroup and profilingOpts to
jitProfilingGroup and jitProfilingOpts to avoid confusion, as they really
are specific to JitProfiling mode.

Signed-off-by:  Henry Zongaro <zongaro@ca.ibm.com>