[Jtreg/FFI]Fix the OOM issue with thunk allocation
The changes aim to address the OOM issue with thunk
allocation on the platforms with a small page size
(e.g. 4KB) by maintaining a thunk heap list in which
case a new thunk heap is created only when there is
no free thunk heap available in the list to allocate
thunk.

Signed-off-by: Cheng Jin <jincheng@ca.ibm.com>