Fix JVMTI ContFramePopTest
If a virtual thread is not mounted and a continuation is being run
directly on a platform thread, then the continuation is considered
part of the platform thread, and the fields should not be swapped
with J9VMContinuation->currentContinuation to get the platform thread
details. This behaviour matches the RI.

Get platform thread details from J9VMThread->currentContinuation IFF
it is not mounted.

Fixes: #16240

Signed-off-by: Babneet Singh <sbabneet@ca.ibm.com>