Do not check for the existence of the AOT cache on critical path
When a first request for a named AOT cache is made, the code would check
for the existence of an appropriately named file cache and then place
a special compilation request to load the cache asynchronously.
However, since this file can be on network attached storage, the
check could take many milliseconds, blocking the current compilation
request and possibly others (due to holding the cache map monitor).
This commit removes the mentioned file check. The only downside is that,
we may delay the creation of an empty in-memory cache and thus we may
not store the first few AOT compilations.

Signed-off-by: Marius Pirvu <mpirvu@ca.ibm.com>