Factor VP interface type signature detection logic into its own method
The definition of the method is copied verbatim from the definition
currently in OMR. Once merged, this copy will allow the Java-specific
logic to be removed from OMR while still using it in OpenJ9.

Factoring the logic into its own method allows for it to be fixed (to
handle array-of-interface types) in a single place.