Make VP erase interface types in array types from signatures
The verifier treats interface types as java/lang/Object even within
array component types. In particular, when any parameter, instance or
static field, or return type is declared with a signature naming an
array of an interface type, its value is guaranteed to be a reference
array, but not necessarily the named array type. For example, if a
parameter is declared to be of type List[], it can actually be any
Object[].