Fix the return type of VectorSupport.fromBitsCoerced
The return type of jdk_internal_vm_vector_VectorSupport_fromBitsCoerced
can be either of Vector or Mask.
Change the type field of the fromBitsCoerced entry in the method table
to unknown and update visitNodeToBuildVectorAliases to get the correct
type for the call by checking the value of the 5th child node.

Signed-off-by: Akira Saitoh <saiaki@jp.ibm.com>