AArch64: Simplify stack offsets for FFI upcall on macOS
This commit partially removes macOS-specific code from FFI upcall for
AArch64.

Linux and macOS use different layout for arguments passed in the stack.
Arguments are always aligned at the 8-byte boundary in the stack on
Linux, while macOS places arguments at 1/2/4/8-byte boundary depending
on their types.
However, the arguments saved by the upcall thunk do not need to follow
the layout when getArgPointer() knows it.

This commit changes the argument layout on macOS to "always at the
8-byte boundary" for the stack extended by the upcall thunk, to simplify
the code.

Signed-off-by: KONNO Kazuhiro <konno@jp.ibm.com>