Fix jvmtiGetThreadInfo to return correct thread group
The jvmtiGetThreadInfo returns the invalid thread group in case of
unstarted threads. A NULL targetThread indicates that the thread is
either unstarted or terminated so this check is not sufficient to
validate terminated threads.

The boolean check is updated in jvmtiGetThreadInfo to return a valid
thread group for unstarted threads.

Fixes: #16226

Signed-off-by: Dipak Bagadiya <dipak.bagadiya@ibm.com>