Fix unsafe call transformation
During the unsafe call transformation, the node that
loads from ramStatics shares the same temp with the node
that loads from the object. The sym ref for ramStatics is
not collected but the sym ref used for object is collected.
Therefore, they cannot share the same temp slots.

Split the blocks to add a block that calls the atomic helper
that loads from ramStatics. The original atomic helper that
loads from object is in another block. The return values from
both helper calls are saved to the same temp.

Fixes: #15251

Signed-off-by: Annabelle Huo <Annabelle.Huo@ibm.com>