Scan virtual thread stacks for code cache reclamation candidates
Code cache reclamation reclaims the memory of old method bodies when
they are recompiled.  The way it proves that an old method body is no
longer used is by hooking the GC exit event and walking all active
thread stacks looking for methods that are candidates for reclaimation.

This approach does not consider virtual threads, and hence thread stacks
for virtual threads that are not presently attached to a J9VMThread must
also be scanned in this manner.

Signed-off-by: Daryl Maier <maier@ca.ibm.com>