Improve client JVM disconnect messages
The messages that are printed by a client JVM when the `-XX:+JITServerLogConnections`
option is specified can sometimes be misleading. For instance, if a client starts,
but there is no server available, the client will print a message saying that it
lost connection to the server, where in fact, it was never connected to a server.
In other cases there could be a transient error that one of the compilation thread
experiences and the client will issue a "Lost connection to the server" message,
but the other compilation threads continue to compile remotely just fine.
This commit fixes the issues mentioned above.

Issue: #16381

Signed-off-by: Marius Pirvu <mpirvu@ca.ibm.com>