Fix String.split when regex is compressed and searched string is not
If compact strings is enabled and the regex and searched strings compression
differs, then accessing the backing array of the regex string will be done
using the searched string's compression level. This gets us bogus values, and
leads to erroneous non match when the regex string is compressed and searched
string is not.
Thus we decompress the regex string in this case.

Signed-off-by: Mike Zhang <mike.h.zhang@ibm.com>