Remove wantToPatchClassPointer() and the hcrPatchClassPointers option
These were introduced a long time ago to allow a certain change to be
toggled. The change was to stop generating runtime assumptions for
patching class pointers on class redefinition, and to stop generating
slower but more easily patchable instruction sequences to allow for
those runtime assumptions. The patching was no longer needed because
class redefinition had been made in-place.

As a result, wantToPatchClassPointer() and hcrPatchClassPointers are no
longer necessary. They are removed, and all use sites now behave as
though hcrPatchClassPointers is not set and wantToPatchClassPointer()
returns false.