Allow allocationFence to be reenabled by Local Flush Elimination
Flow Sensitive Escape Analysis will mark an allocationFence with
omitSync set to true if it was needed by a single allocation that ended
up being stack allocated.

On a subsequent pass of Escape Analysis, Local Flush Elimination might
decide an allocationFence is needed at the same spot for another object
allocation.  An assertion in TR_LocalFlushElimination::examineNode
expects any existing allocationFence that it finds to have its omitSync
set to false, which didn't account for the possibility that it might
find one that had previously been disabled.

Fixed the problem by removing the assertion and allowing the
allocationFence to be reactivated for the second object allocation.

Also added more information to some trace messages.

Signed-off-by:  Henry Zongaro <zongaro@ca.ibm.com>