Look for interface call targets using the known non-null interface
In f544293aaa4c38ef findCallSiteTarget() was updated to ensure that the
interface is found before attempting findCallSiteTargetImpl(). This is
necessary because getClassFromMethod() can return null in AOT. However,
the resulting interface was used only for assertions that run after
successfully finding a target, and findCallSiteTargetImpl() would repeat
getClassFromMethod() to find the interface again, creating a second
opportunity to fail and get null in AOT.

With this commit, the interface obtained by findCallSiteTarget() is now
reused when actually selecting a target, since it's known to be non-null
by the time we reach findCallSiteTargetImpl().