Fix ICF and register dependencies in CRC-32C acceleration on Z
The first branch to OOL of the acceleration implementation and the
merge point from OOL were not included in the ICF, and the correct
dependencies were not attached to the merge point. This opened up the
possibility of register spills being inserted that would be missed if
the first branch to OOL is taken. This patch resolves this issue by
extending the ICF from the first branch to OOL to the merge point and
moving all the dependencies to the merge point.

Fixes: #16560
Signed-off-by: Spencer Comin <spencer.comin@ibm.com>