Only accept -Xenablefips140-3 on supported platforms
The FIPS 140-3 tech preview is currently only available on 64-bit
X86 and POWER platforms, so only accept the -Xenablefips140-3 (and
-Xdisablefips140-3) options on those platforms. On 64-bit Z for Linux,
the command-line options should be rejected but the system properties
need to be set (because the preview directory structure is in place).
On other platforms, the options will be rejected and the
com.ibm.fips.home and com.ibm.fips.mode properties will NOT be set for
consistency with non-J9 platforms (which also do not have the tech
preview). Also only make changes to the extended classpath directory
on the 5 platforms identified above where the system properties are set.

Defines FIPS_PREVIEW_OPTIONS_ACCEPTED_PLATFORM on platforms where the
command-line options should be processed and where the preview can be
enabled. Also defines FIPS_PREVIEW_PLATFORM on platforms where the
directory structure has been modified and so the system properties
need to be set and the classpaths need to be updated. Eventually these
two macros will be set on the same set of platforms, but not yet.

Signed-off-by: Mark Stoodley <mstoodle@ca.ibm.com>