Unbounded GC Thread Pool Expansion for CRIU Restore
Initial set of changes to expand the GC thread pool beyond the
dispatcher initialization done at VM startup. These changes are required
by the consumed OMR GC dispatcher/configuration APIs called for restore.

- Converted reference obj list iteration count from `gcThreadCount` to
`regionExtension->_maxListIndex`, this is more correct. This is the only
instance in the code where iteration implies the listCount from GC
thread count. This inference was not an issue as the lists are inited
based on the GC thread count, however now that the thread count may
change, it is incorrect to assume the list count from the thread count.
- Release restore thread's VM access for the duration of
`reinitializeGCThreadCountForRestore`

Signed-off-by: Salman Rana <salman.rana@ibm.com>