Track JITServer thunks for AOT methods
Compilations that are to be stored in the JITServer AOT cache now handle
thunks differently from non-AOT compilations; instead of maintaining the
per-client thunk pointer map directly, we now generate thunks at the
server for those compilations and store them in the JITServer AOT cache.
The compilation object also tracks any thunks a compilation might
require, so the relevant records can be added to the main
CachedAOTMethod record and be sent to clients as needed.

This separate handling is required to ensure that clients have in their
local shared class cache any thunks that a JITServer AOT cache method
might require. Not having these thunks available will cause an AOT
load failure because of the resulting thunk relocation error.

Fixes: #16482
Signed-off-by: Christian Despres <despresc@ibm.com>