Allow for overriding Simplifier handlers defined by OMR
OMR pull request https://github.com/eclipse/omr/pull/6821 introduces
changes to the Simplifier optimization that allows a downstream project
to override handlers that are supplied for core OMR opcodes in addition
to supplying handlers for opcodes that the downstream project provides.
The overriding is performed in the new files SimplifierHandlers.hpp and
SimplifierTable.enum.

This is a coordinated change that allows OpenJ9 to accommodate the OMR
changes without actually taking advantage of the ability to override
any simplifier handlers.

Signed-off-by:  Henry Zongaro <zongaro@ca.ibm.com>