Address vlog/rtLog post restore
Both the vlog and rtlog have the same semantics. If the vlog/rtlog are
specified pre-checkpoint and not post-restore, they remain unchanged. If
the vlog/rtlog is NOT specified pre-checkpoint but IS specified
post-restore, then the vlog/rtlog is opened post-restore. If they are
specified both pre-checkpoint and post-restore, then the pre-checkpoint
file(s) are closed and the post-restore file(s) are opened. A
consequence of this is that if the same file name is specified then a
new file will be opened. This can have the consequence of overwriting
the previous file if the PID of the restored process is the same.

Signed-off-by: Irwin D'Souza <dsouzai.gh@gmail.com>