Support for restore args env var
This PR builds on the existing support for EnvVars
(see #13545) by
recognizing a special env var OPENJ9_RESTORE_JAVA_OPTIONS which can be
used to specify JVM options at upon JVM restore. This env var is ignored
when -XX:+EnableCRIUSupport is not specified and it is also ignored at
startup.

Documentation will be added later to describe what kind of options can
be specified upon restore. Initially all options will be parsed, but this PR
only supports the addition of Java properties, all other options will
be ignored.

All options are placed in the vm->checkpointState.restoreArgsList, this
is what the JVM components (VM, JIT, GC) will look at to extract restore
options.

This PR parses the JVM options before any Java code is run upon restore
so that the JVM can act on the changes immediately. A tracepoint is added
to display all the restore options (-Xtrace:print=j9criu.16), further
support will be added to javadump.

A restore option "-XX:+IgnoreUnrecognizedRestoreOptions" is added, this
option notifies the JVM to ignore any options that are not supported on
restore. Without this option the JVM will throw a restore exception if
an unsupported option is specified.

Related: #15844
Closes: #15839

Signed-off-by: Tobi Ajila <atobia@ca.ibm.com>