Test result of arraycmp for value type comparison
The arraycmp operation produces a result of zero if the memory being
compared is equal, one if the first operand is less than second, or two
if the second operand is less than the first.  The transformation of
<object{Equality|Inequality}Comparison> used arraycmp assuming that the
result would be zero for equality or one for inequality.  This change
adds an icmpeq or icmpne to compare the result of the arraycmp with zero
in order to yield a value in the expected range:  {0,1}.

Signed-off-by:  Henry Zongaro <zongaro@ca.ibm.com>