Update user java.ext.dirs with VM's fips-specific ext directory
When we created a second "jre/lib/ext" directory for the FIPS
140-3 technology preview, we moved some of the files in jre/lib/ext
into separate directory. We automatically set up the java.ext.dirs
system property to include these two directories now. But if the
user overrides it by using -Djava.ext.dirs=<something else> then
there is a strong chance that they won't add the second directory
we created and the Java security implementation may not work
properly in this case. To avoid the requirement that every user
must update any uses of -Djava.ext.dirs=<something>, we instead
intercept such user overrides and automatically append the new
directory. While it is unfortunate that we have to modify the
option setting from the user, it avoids the far more painful
breakage that would otherwise occur.

Since this change is made late in the SR8 release cycle, the
change is intentionally made to happen only if the user overrides
the java.ext.dirs system property so that our existing test results
do not need to be repeated. If no java.ext.dirs option is set by
the user, the paths initialized by the VM should be used without
alteration.

Signed-off-by: Mark Stoodley <mstoodle@ca.ibm.com>