Track EDO compilations
Added `HasEdoSnippet` flag in `TR_PersistentJittedBodyInfo` which is set
when the codegen generates an EDO recompilation snippet (such snippet
decrements `TR_PersistentJittedBodyInfo._counter` whenever a catch
block is entered).
Added new reason for recompilation in `TR_PersistentMethodInfo` called
`RecompDueToEdo` which is set when a recompilation is triggered from
the snippet mentioned above.
Verbose log will show "E" (from EDO) as the reason for recompilation
when the recompilation was triggered by EDO.

Signed-off-by: Marius Pirvu <mpirvu@ca.ibm.com>