Added _catchBlockCounter to TR_PersistentMethodInfo
This counter is supposed to be incremented by jitted code
every time a catch block is entered. The value of the
counter should be inspected by the Inliner to determine
whether it should inline more aggressively on the throw
path in order to transform a throw into a goto operation.

Signed-off-by: Marius Pirvu <mpirvu@ca.ibm.com>