Pin virtual threads in JVMTI RawMonitorEnter and RawMonitorExit
JVMTI functions RawMonitorEnter and RawMonitorExit can be invoked
from Java code using native methods.

ownedMonitorCount is incremented for the current J9VMThread in
RawMonitorEnter after successfully acquiring the monitor, and
it is decremented for the current J9VMThread in RawMonitorExit
after successfully releasing the monitor.

This prevents virtual threads to yield if it owns a monitor.

Related: #16753

Signed-off-by: Babneet Singh <sbabneet@ca.ibm.com>