Ensure jit tracing occurs post restore
If logging isn't enabled, then the log file is suppressed via
TR::Options::suppressLogFileBecauseDebugObjectNotCreated. This needs to
be reset post-reset if logging is enabled.

Additionally, another scenario where it needs to be reset is when a
remote compilation request fails. In non-portable CRIU mode, the ability
to connect to a JITServer is implicitly enabled. However, when logging
is specified, the client suppresses the log as the server will generate
the log. However, if the compilation fails, the flag to suppress the log
file should be reset if the compilation will be performed locally.

Signed-off-by: Irwin D'Souza <dsouzai.gh@gmail.com>