Ensure methods can be excluded post restore
To prevent a compiled body from running ever again (e.g., because it was
excluded in the post-restore option), the method has to be invalidated.
This may trigger a recompilation, which will be denied because of the
exclude filter. This results in a call to
J9::Recompilation::methodCannotBeRecompiled which should patch the Start
PC to call a helper to transition to the interpreter.

This commit updates the logic that hat determines whether to trasition
to the interpreter or continue executing the existing method body had
to be augmented to include methods invalidated because they were
excluded post-restore.

Signed-off-by: Irwin D'Souza <dsouzai.gh@gmail.com>