Reset start and elapsed time
The Checkpoint phase is conceptually part of building the application;
therefore it does not make sense to expect a user who specifies an
option such as `-XsamplingExpirationTime` to take into account the time
spent executing in the Checkpoint phase. Therefore, on restore, both
the start and elapsed time are reset.

Signed-off-by: Irwin D'Souza <dsouzai.gh@gmail.com>