Enable com.ibm.fips.mode and .home to be set on command-line
There are scenarios (e.g. when ant forks a JVM to run jobs) where
the com.ibm.fips.mode and .home properties are set explicitly on
the command-line (via -D options). A problem arises because the VM
also internally caches fipsHome and fipsMode (when they are
initialized by default by the VM) and uses those values later to do
some of its internal configuration. When those system properties
are explicitly set on the command line, these internally cached
fipsHome and fipsMode values are not properly initialized. Then
later they are read, and extreme disappointment ensues.

There are already code paths in the VM that test if these properties
are set; but these "if" statements currently don’t have “else” clauses.
The fix is to add “else” clauses that do some validation on the values
that have been provided and then store the values into those internal
caches so that everything can proceed as originally intended.

Signed-off-by: Mark Stoodley <mstoodle@ca.ibm.com>