Fixing wrong address elementSize calculation
generateReportFieldAccessOutlinedInstructions incorrectly believes that when it
is run on a awrtbar/awrtbari node and compressed refs is active, it is being
passed a compressed object address to pass into a call to the VM. In reality,
it is being given an uncompressed object address and the VM is expecting an
uncompressed value. This causes a problem where a stw instruction is used to
store a 64 bit value.

This change removes the special case handling of a compressed ref value for
calculating elementSize. By default, convertTypeToSize already returns the
correct value of 8 for 64 bit (and 4 for 32 bit).

Issue: #16750
Signed-off-by: jimmyk <jimmyk@ca.ibm.com>