Pass valid JNI refs to getVirtualThreadState
(jthread)&threadObject is not a JNI ref, and it will not be
updated by the GC when VM access is released in
jvmtiHelpers.c::getVirtualThreadState.

Non-null thread refs are directly passed to getVirtualThreadState.
A local ref is created for currentThread->threadObject before it is
passed to getVirtualThreadState. This assures that the GC updates
the virtual thread object which is passed to getVirtualThreadState.

In getVirtualThreadState, the virtual thread object is refected
after VM access is re-acquired to prevent access to a stale object.

Signed-off-by: Babneet Singh <sbabneet@ca.ibm.com>