Fix memory leak in fips.mode and fips.home handling
When a user tries to directl override the values of com.ibm.fips.mode
or com.ibm.fips.home, the VM allocates memory to copy the values used,
but it never frees this memory.

In the case of com.ibm.fips.mode, there is actually no need to copy
the value because we ensure it can only be set to "140-2" or "140-3"
so we can just use these literal strings. In the case of com.ibm.fips.home,
we do not know what value will be used, so we copy it. We remember
that we made a copy in this case and then free the memory when
libraries are being unloaded.

Signed-off-by: Mark Stoodley <mstoodle@ca.ibm.com>