Add support for -Xrs on restore
This PR adds support for -Xrs on restore. Given that its imposible to
support -Xrs in the same manner on restore that it is supported at startup
(because there may be frames on the stack with signal handlers already
installed that cannot be removed on restore) this change requires that
"onRestore" is appended to the option.

For example: -Xrs:sync -> -Xrs:syncOnRestore, -Xrs -> -Xrs:onRestore, ..

Signed-off-by: Tobi Ajila <atobia@ca.ibm.com>