Delete TR_J9VMBase::getFieldOffset(), which is now unused
Previously the only call site was in ilgen when transforming recognized
methods of com/ibm/Compiler/Internal/Prefetch, which it no longer does.

With this deletion, the owning method and constant pool index of a const
string symref should only matter if it's unresolved. That will make it
possible to mark arbitrary (resolved) known object const refs as const
strings (whenever they happen to be strings).